requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Segmentation Models
segmentation-models-pytorch>=0.3.3

# Data Augmentation
albumentations>=1.3.0

# Data Handling
numpy>=1.24.0
Pillow>=9.5.0

# Visualization & Progress
matplotlib>=3.7.0
tqdm>=4.65.0

# Optional: Experiment Tracking (uncomment if needed)
# wandb>=0.15.0
# tensorboard>=2.13.0
